# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cross App Backend",
    "description": "API para gestao e gamificacao de CrossFit",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
...
    "/feed/upload": {
      "post": {
        "tags": [
          "Feed"
        ],
        "summary": "Upload de imagem para post do feed",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Imagem invalida ou ausente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para postar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/feed/post": {
      "post": {
        "tags": [
          "Feed"
        ],
        "summary": "Cria post no feed do box",
        "description": "Recebe texto + foto opcional e vincula o post a um check-in valido do aluno. Um check-in aceita somente um post.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateFeedPostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos ou check-in nao encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para postar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Ja existe post para o check-in informado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/results": {
      "post": {
        "tags": [
          "Results"
        ],
        "summary": "Registra resultado de treino e identifica novo PR",
        "description": "Salva o score do aluno para um exercicio em um WOD e marca isNewPR=true quando o score supera o melhor historico do proprio aluno para o mesmo exercicio.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResultDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resultado salvo com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para registrar resultado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "WOD ou exercicio nao encontrado para o box",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/results/pr": {
      "post": {
        "tags": [
          "Results"
        ],
        "summary": "Registra PR direto por exercicio",
        "description": "Permite ao aluno selecionar um exercicio (via GET /exercises) e salvar um PR sem vincular ao WOD, com comparacao de historico por exerciseId.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExercisePrDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PR salvo com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResultResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para registrar PR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Exercicio nao encontrado para o box",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/students": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Lista todos os alunos do box do usuario autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de alunos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me/enrollment-token": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Gera token temporario do aluno para matricula em um box",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Token gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTokenResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para gerar token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/enroll": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Vincula um aluno ao box atual usando o token temporario do aluno",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UseEnrollmentTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Aluno vinculado ao box com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollStudentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Token invalido, expirado ou ja utilizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para matricular aluno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Aluno ja matriculado neste box",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "XBoxIdHeader": {
        "name": "x-box-id",
        "in": "header",
        "required": true,
        "description": "ID do box selecionado",
        "schema": {
          "type": "string",
          "example": "67ea76a5ac5d89c8bb9d2111"
        }
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "string",
        "pattern": "^[a-fA-F0-9]{24}$",
        "example": "67ea76a5ac5d89c8bb9d2111"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "COACH",
          "ALUNO"
        ]
      },
...
      "ResultScoreKind": {
        "type": "string",
        "enum": [
          "TIME",
          "LOAD",
          "UNKNOWN"
        ]
      },
...
      "CreateFeedPostDto": {
        "type": "object",
        "required": [
          "checkinId",
          "text"
        ],
        "properties": {
          "checkinId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "text": {
            "type": "string",
            "minLength": 2,
            "maxLength": 1200,
            "example": "Aula de hoje foi insana. Novo PR no deadlift!"
          },
          "photoUrl": {
            "type": "string",
            "format": "uri",
            "example": "http://localhost:3000/uploads/feed/1712086421000-123456789.jpg"
          }
        }
      },
...
      "CreateResultDto": {
        "type": "object",
        "required": [
          "wodId",
          "exerciseId",
          "score"
        ],
        "properties": {
          "wodId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "exerciseId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "score": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "95kg",
            "description": "Score do resultado. Aceita carga (ex: 95, 95kg) ou tempo (ex: 02:34, 00:12:30)."
          },
          "autoPostText": {
            "type": "string",
            "minLength": 2,
            "maxLength": 1200,
            "example": "Bati PR hoje no WOD! Muito feliz com a evolucao.",
            "description": "Texto customizado para auto-post no feed quando houver novo PR"
          }
        }
      },
      "CreateExercisePrDto": {
        "type": "object",
        "required": [
          "exerciseId",
          "score"
        ],
        "properties": {
          "exerciseId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "score": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "example": "100kg",
            "description": "Score do PR. Aceita carga (ex: 95, 95kg) ou tempo (ex: 02:34, 00:12:30)."
          },
          "autoPostText": {
            "type": "string",
            "minLength": 2,
            "maxLength": 1200,
            "example": "PR no squat! Consistencia ta valendo a pena.",
            "description": "Texto customizado para auto-post no feed quando houver novo PR"
          }
        }
      },
...
      "FeedUploadResponse": {
        "type": "object",
        "required": [
          "url",
          "filename",
          "message"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "http://localhost:3000/uploads/feed/1712086421000-123456789.jpg"
          },
          "filename": {
            "type": "string",
            "example": "1712086421000-123456789.jpg"
          },
          "message": {
            "type": "string",
            "example": "Upload realizado com sucesso"
          }
        }
      },
      "CreateFeedPostResponse": {
        "type": "object",
        "required": [
          "postId",
          "checkinId",
          "message"
        ],
        "properties": {
          "postId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "checkinId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "message": {
            "type": "string",
            "example": "Post criado com sucesso"
          }
        }
      },
...
      "Student": {
        "type": "object",
        "required": [
          "_id",
          "boxIds",
          "name",
          "email",
          "role",
          "createdAt"
        ],
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "boxIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "name": {
            "type": "string",
            "example": "Joao Aluno"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@aluno.com"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateResultResponse": {
        "type": "object",
        "required": [
          "resultId",
          "isNewPR",
          "scoreKind",
          "autoFeedPost",
          "message"
        ],
        "properties": {
          "resultId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "isNewPR": {
            "type": "boolean",
            "example": true
          },
          "scoreKind": {
            "$ref": "#/components/schemas/ResultScoreKind"
          },
          "autoFeedPost": {
            "$ref": "#/components/schemas/AutoFeedPostResult"
          },
          "message": {
            "type": "string",
            "example": "Resultado salvo com sucesso"
          }
        }
      },
      "AutoFeedPostResult": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "created",
              "skipped-no-checkin",
              "skipped-already-posted",
              "skipped-no-new-pr"
            ]
          },
          "postId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "checkinId": {
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "statusCode": {
            "type": "number",
            "example": 400
          }
        }
      },
...
      "UseEnrollmentTokenDto": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{32}$",
            "minLength": 32,
            "maxLength": 32,
            "example": "2a90e8f904194ea2c1435d85b8d628a2"
          }
        }
      },
      "EnrollmentTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "expiresAt",
          "message"
        ],
        "properties": {
          "token": {
            "type": "string",
            "example": "2a90e8f904194ea2c1435d85b8d628a2"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "example": "Token gerado com sucesso. Informe este codigo ao administrador do box."
          }
        }
      },
      "EnrollStudentResponse": {
        "type": "object",
        "required": [
          "studentId",
          "boxId",
          "name",
          "email",
          "role",
          "message"
        ],
        "properties": {
          "studentId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "boxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "example": "Joao Aluno"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@aluno.com"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "message": {
            "type": "string",
            "example": "Aluno vinculado ao box com sucesso"
          }
        }
      },
...
    }
  }
}